harvest_link/requirements.txt
google-adk>=0.1.0
google-genai>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
